Properly detect missing SMART support
Reported by Patrick Kuijvenhoven on https://github.com/ercpe/smart-check/issues/5.

Previously, we did not check the value of `SMART support is` in the information section of the smartctl output and relied entirely on the other sections.
Now we check that value too and report a failed check if the device (or maybe a raid controller) announces that the device has no SMART capability.